Move types to composite-checkout dir
