+++ client/my-sites/checkout/composite-checkout/types/index.ts
