+++ client/my-sites/checkout/composite-checkout/types/transaction-endpoint.ts
